Cleaned Up a little bit it is defienetly more readable hehehe
