+++ builtins/cd_com.c
@@ -1,0 +1,48 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   cd_com.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aruckenb <aruckenb@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/10/18 11:35:39 by aruckenb          #+#    #+#             */
+/*   Updated: 2024/10/18 11:36:22 by aruckenb         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../minishell.h"
+
+void	cd_com(t_data *core)
+{
+	char *old_pwd;
+
+	old_pwd = getcwd(NULL, 0);
+	if (ft_strncmp(core->line, "cd -", 4) == 0)
+	{
+		int i = 0;
+		while (core->env[i])
+		{
+			if (ft_strncmp(core->env[i], "OLDPWD=", 7) == 0)
+			{
+				free(core->direct);
+				core->direct = ft_strdup(core->env[i] + 7);
+				break ;
+			}
+			i++;
+		}
+		ft_printf("%s\n", core->direct);
+		chdir(core->direct);
+		envi_update(old_pwd, core);
+		return ;
+	}
+	free(core->direct);
+	core->direct = ft_strdup(core->line + 3);
+	if (access(core->direct, sizeof(char)) == 0)
+	{
+		chdir(core->direct);
+		envi_update(old_pwd, core);
+	}
+	else
+		ft_printf("cd: %s: no such file or directory\n", core->direct);
+	return ;
+}
